Added Enter key support. Added error prevention when clicking "Analyze app"
The enter key now simulation pressing "Analyze app" while the apps table view is currently selected. If the apps table view is not selected, then nothing will happen.

When "Analyze app" clicked, it now checks to see if the user has selected an app and presents an NSAlert if no app is selected.