Renamed button/method and added default environment varibales to NSTask
Renamed the "Build/Load Tweak" button to "Install Tweak". NSTask when creating a new theos project relied on the $THEOS environment variable but when I launched the task it would not import that variable. I fixed this by adding the argument "-l" (that's an L) to the array of arguments passed to the setArguments method.