Checks to see is Application Support symbolic link exists
Checks to see is Application Support symbolic link exists as Application-Support, if not then ~/Library/Application-Support is created as an alias to "~/Library/Application Support". The "-" is needed because when theos tries to build a package, the path of the package cannot have a space in the name. Also AMFileManager's mainDirectoryPath property now uses the symlink.