UI changed and clear log button additional functionality
I changed the UI so that it is nicer when the user is resizing the window, before all the buttons and views wouldn't scale/move nicely. The clear log button used to only clear the text in the iOS application log view but now it actually removes the log file (AMLog.txt) on the target app on the iOS device as well. AMLogger's logForApp:app method has been renamed to retrieveLogForApp:app. AMLogger has a new method called - removeLogForApp:app which deletes the AMLog.txt stored in the selected app's data directory on the iOS device.