Frida now fully supported
I had to make my own fork of frida-ios-dump because I needed to be able to pass more command line arguments as well as add the option to only dump the decrypted binary instead of the whole .ipa. Making the fork of frida work with binary only mode was such a challenge but after a few hours (maybe 6?) I finally got it to work!