Back to static libraries
Actually we're back to using static libraries! I am trying to minimize the amount of external dependencies so I am now bundling liplist.a and libimobiledevice.a in a new folder called Libraries. There is still an issue I encountered when using this on a machine that wasn't mine and it has to do with lockdown. I will attempt to figure this out soon but we're taking baby steps. I think once I figure out the lockdown issue I'll be ready to ship externally!