Fixed problem when logging methods that return primitives
If a method is not void then I normally log the a pointer to the object returned by the method as well as the value of the object. This does not work if the methods returns a primitive (e.g. unsigned long long). Now, if a method returns a primitive we no longer log the pointer, just the return value.