Added version of app to AMApp
We will use this in the future to determine if we need to to decrypt an app again. If the app that has been decrypted has the same version the app on the device, then there is no need. If the versions differ, then decrypt the binary on the device.